+++ src/Ext/TAction/Body.h
@@ -11,7 +11,10 @@ class HouseClass;
 
 enum class PhobosTriggerAction : unsigned int
 {
-	SaveGame = 500
+	SaveGame = 500,
+	EditVariable,
+	GenerateRandomNumber,
+	PrintVariableValue,
 };
 
 class TActionExt
@@ -44,6 +47,15 @@ public:
 			ObjectClass* pObject, TriggerClass* pTrigger, CellStruct const& location);
 
 	static bool SaveGame(TActionClass* pThis, HouseClass* pHouse,
+			ObjectClass* pObject, TriggerClass* pTrigger, CellStruct const& location);
+
+	static bool EditVariable(TActionClass* pThis, HouseClass* pHouse,
+			ObjectClass* pObject, TriggerClass* pTrigger, CellStruct const& location);
+
+	static bool GenerateRandomNumber(TActionClass* pThis, HouseClass* pHouse,
+			ObjectClass* pObject, TriggerClass* pTrigger, CellStruct const& location);
+
+	static bool PrintVariableValue(TActionClass* pThis, HouseClass* pHouse,
 			ObjectClass* pObject, TriggerClass* pTrigger, CellStruct const& location);
 
 	class ExtContainer final : public Container<TActionExt>